backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==1.10.13
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.30.1
